Fixed Integration Test for Component and Route property binding (#5)
<!-- Provide a general summary of your changes in the Title above -->
<!-- Apply the label "bug" or "enhacement" as applicable. -->

## Description / Motivation
<!-- Describe your changes in detail -->
<!-- Why is this change required? What problem does it solve? -->
<!-- If it fixes an open issue, please link to the issue here. -->

The integration test for Component and Route property binding mistakenly
had `BeEmpty()` instead of `Be()`. This incorrectly made the integration
test pass since the Razor View was invalid. It was using the `asp-for`
TagHelper attribute to attempt to render the values. However the Model's
properties are of type `string?` which do not have an OOTB TagHelper for
the `h3` or `p` tags. With the view fixed this test now properly passes
and behaves as expected.

## Testing

- [X] The Unit & Intergration tests are passing.
- [X] I have added the necesary tests to cover my changes.

## Terms
<!-- Place an X in the [] to check. -->

<!-- The Code of Conduct helps create a safe space for everyone. We
require that everyone agrees to it. -->
- [X] I agree to follow this project's [Code of
Conduct](CODE_OF_CONDUCT.md).

Co-authored-by: Ivan Lieckens <ivanlieckens@hotmail.com>