Fixed issue rendering nested empty placeholders (#6)
Currently, empty Placeholders are not rendering the same HTML that JSS
does. This PR updates the SDK's TagHelper to generate HTML that matches
the same output from JSS.

## Description / Motivation
When a Placeholder is rendered without any components loaded inside of
it, it needs to be wrapped in a div to be correctly rendered in Pages
and Experience Editor. The div being created by for this SDK is `<div
className=\"sc-empty-placeholder\"></div>`

## Testing
- [x] The Unit & Intergration tests are passing.
- [x] I have added the necesary tests to cover my changes.

## Terms
- [x] I agree to follow this project's [Code of
Conduct](CODE_OF_CONDUCT.md).

---------

Co-authored-by: Ivan Lieckens <ivanlieckens@hotmail.com>