Handle empty string fields in FieldConverter. (#12)
<!-- Provide a general summary of your changes in the Title above -->
<!-- Apply the label "bug" or "enhacement" as applicable. -->

## Description
This update addresses a specific issue encountered when dealing with
JSON data for fields, such as the following example:
```
"fields": {
    "title": "",
    "leftLink": {
        "key": "40dfd6adb1334ff48c3e4c68ab9b9b35",
        "text": "The Energy Transition",
        "href": "/connections/energy-transition-and-sustainability/what-is-the-energy-transition"
    },
    "rightLink": {
        "key": "de61e12028784b4da96a9676b73fc764",
        "text": "There are currently 54 known outages.",
        "href": "/outages-and-maintenance"
    }
}

```
Previously, an empty string in the `title` field would trigger a
`JsonException` in the `FieldConverter` because it was not recognized as
a valid `JsonValueKind` type. This update introduces support for
`JsonValueKind.String`, ensuring that empty strings are handled
correctly alongside objects and arrays, thereby resolving the error.

## Motivation
Custom Content Resolvers in both XM Cloud and Sitecore XP may return
empty strings for certain fields, leading to errors in the
`FieldConverter`. This enhancement ensures that such cases are handled
properly, allowing for smooth operation regardless of whether an empty
string is encountered.

**Update:** Following the discussion below, this PR has been revised to
focus on adding an integration test for fields with empty strings
instead of altering the `FieldConverter`.