Context Support (#18)
<!-- Provide a general summary of your changes in the Title above -->
<!-- Apply the label "bug" or "enhacement" as applicable. -->

## Description / Motivation
<!-- Describe your changes in detail -->
This allows creating connections to the Edge using the Edge Platform
Context instead of an ApiKey. It introduces new extension methods on the
`SitecoreLayoutClientBuilder` called `AddGraphQlWithContextHandler` for
the LayoutService.Client and enhances the extensions on
`GraphQlConfiguration` called `AddGraphQlClient` for the GraphQL.
Introducing `ContextId` to the `SitecoreGraphQlClientOptions` as well as
by default using the
https://edge-platform.sitecorecloud.io/v1/content/api/graphql/v1
endpoint instead of throwing an exception when passing in a ContextId.

<!-- Why is this change required? What problem does it solve? -->
Moving forward Sitecore and XM Cloud in particular will greatly rely on
the unified Edge Platform endpoint with
[ContextId](https://doc.sitecore.com/xmc/en/developers/xm-cloud/the-context-id-environment-variable.html)
making this an important addition to the SDK.

<!-- If it fixes an open issue, please link to the issue here. -->
This resolves #10 

## Testing

- [X] The Unit & Intergration tests are passing.
- [X] I have added the necessary tests to cover my changes.

## Terms
<!-- Place an X in the [] to check. -->

<!-- The Code of Conduct helps create a safe space for everyone. We
require that everyone agrees to it. -->
- [X] I agree to follow this project's [Code of
Conduct](CODE_OF_CONDUCT.md).

---------

Co-authored-by: Ivan Lieckens <ivanlieckens@hotmail.com>