Changed text based TagHelpers to rendering Chromes with internal span not div (#40)
This fixes the bug where Text based TagHelpers can't render inside of a
`p` tag, causing all of MetaData editing to break.

This aligns with the JSS approach of using a `span` tag instead.

I also introduced a new `Assert` to the appropriate UnitTests as that
tag insertion wasn't being tested currently.

## Testing

- [x] The Unit & Intergration tests are passing.
- [x] I have added the necessary tests to cover my changes.

## Terms
<!-- The Code of Conduct helps create a safe space for everyone. We
require that everyone agrees to it. -->
- [x] I agree to follow this project's [Code of
Conduct](CODE_OF_CONDUCT.md).

Closes #39